refactor : Member 관련 API 분해
